Move DecryptedRecord definitions into NetworModelContract
