+++ sdk-core/src/main/java/care/data4life/sdk/model/ModelContract.kt
@@ -18,7 +18,7 @@ package care.data4life.sdk.model
 
 import care.data4life.fhir.stu3.model.DomainResource
 import care.data4life.sdk.lang.CoreRuntimeException
-import care.data4life.sdk.network.model.definitions.DecryptedBaseRecord
+import care.data4life.sdk.network.model.NetworkModelContract.DecryptedBaseRecord
 import care.data4life.sdk.tag.Annotations
 import org.threeten.bp.LocalDate
 import org.threeten.bp.LocalDateTime
