+++ sdk-core/src/main/java/care/data4life/sdk/network/model/DecryptedR4Record.kt
@@ -18,7 +18,7 @@ package care.data4life.sdk.network.model
 
 import care.data4life.crypto.GCKey
 import care.data4life.sdk.fhir.Fhir4Resource
-import care.data4life.sdk.network.model.definitions.DecryptedFhir4Record
+import care.data4life.sdk.network.model.NetworkModelContract.DecryptedFhir4Record
 import care.data4life.sdk.tag.Annotations
 import care.data4life.sdk.tag.Tags
