+++ sdk-core/src/test/java/care/data4life/sdk/model/RecordMapperTest.kt
@@ -28,7 +28,7 @@ import care.data4life.sdk.model.ModelContract.RecordFactory
 import care.data4life.sdk.network.model.DecryptedDataRecord
 import care.data4life.sdk.network.model.DecryptedR4Record
 import care.data4life.sdk.network.model.DecryptedRecord
-import care.data4life.sdk.network.model.definitions.DecryptedBaseRecord
+import care.data4life.sdk.network.model.NetworkModelContract.DecryptedBaseRecord
 import care.data4life.sdk.tag.Annotations
 import care.data4life.sdk.tag.Tags
 import care.data4life.sdk.wrapper.SdkDateTimeFormatter
@@ -100,7 +100,7 @@ class RecordMapperTest {
         val record = RecordMapper.getInstance(decryptedRecord)
 
         // Then
-        assertTrue(record is Fhir3Record)
+        assertTrue(record is Fhir3Record<*>)
         assertEquals(
             actual = record.identifier,
             expected = ""
