+++ sdk-core/src/test/java/care/data4life/sdk/network/DecryptedRecordMapperTest.kt
@@ -24,7 +24,8 @@ import care.data4life.sdk.network.model.DecryptedR4Record
 import care.data4life.sdk.network.model.DecryptedRecord
 import care.data4life.sdk.network.model.DecryptedRecordGuard
 import care.data4life.sdk.network.model.NetworkModelContract
-import care.data4life.sdk.network.model.definitions.DecryptedFhir3Record
+import care.data4life.sdk.network.model.NetworkModelContract.DecryptedCustomDataRecord
+import care.data4life.sdk.network.model.NetworkModelContract.DecryptedFhir3Record
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
@@ -171,7 +172,7 @@ class DecryptedRecordMapperTest : DecryptedRecordMapperTestBase() {
         )
 
         // Then
-        assertTrue(record is care.data4life.sdk.network.model.definitions.DecryptedCustomDataRecord)
+        assertTrue(record is DecryptedCustomDataRecord)
         assertEquals(
             record,
             care.data4life.sdk.network.model.DecryptedDataRecord(
